LayoutTests/media/video-seek-with-negative-playback-expected.txt


Test that seeking with a negative playbackRate works correctly.

EVENT(canplaythrough)
EVENT(seeked)
EVENT(timeupdate)
EVENT(timeupdate)
EXPECTED (video.currentTime < '3') OK
END OF TEST
